+++ index.html
@@ -10,7 +10,7 @@
 		<script type="text/javascript" src="/lib/litegraph.core.js"></script>
 		<script type="text/javascript" src="/lib/litegraph.extensions.js" defer></script>
 		<script type="module">
-			import { app } from "./src/scripts/app.js";
+			import { app } from "./src/scripts/app";
 			(async () => {
 				await app.setup();
 				window.app = app;
