+++ src/scripts/ui.ts
@@ -2,7 +2,7 @@ import { api } from "./api";
 import { ComfyDialog as _ComfyDialog } from "./ui/dialog.js";
 import { toggleSwitch } from "./ui/toggleSwitch.js";
 import { ComfySettingsDialog } from "./ui/settings.js";
-import { ComfyApp, app } from "./app.js";
+import { ComfyApp, app } from "./app";
 
 export const ComfyDialog = _ComfyDialog;
